requirements.txt
Flask
pyspellchecker
graphviz